requirements.txt
# Core Flask Dependencies
Flask>=2.3.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.0
SQLAlchemy>=2.0.0
Werkzeug>=2.3.0

# Database
psycopg2-binary>=2.9.0

# Authentication & Security
Flask-Bcrypt>=1.0.1

# Image Processing
Pillow>=10.0.0

# Google Gemini AI
google-generativeai>=0.3.0

# Firebase/Firestore (Optional - for Infrastructure Projects ledger)
firebase-admin>=6.2.0

# Utilities
python-dotenv>=1.0.0

# Date/Time
python-dateutil>=2.8.2

beautifulsoup4
opencv-python-headless
scikit-learn

Flask-WTF

email-validator


